+++ ML_multi_DownhillSimplex.h
@@ -124,7 +124,7 @@ inline void ML_multi_DownhillSimplex::maximize(const v_ratep_type& lb,
     assert(lb.size() == ub.size());
     int num_ratep = branch_rate_manager.get_ratepvector().size();
     double delta = determine_machine_delta();
-    assert(num_ratep == lb.size());
+    assert(num_ratep == (int)lb.size());
     assert(num_ratep > 1);
     set_function_tolerance(f_tol);
     set_scale_length(s_len);
